update for python 2.5
git-svn-id: http://svn.berlios.de/svnroot/repos/python-silc/trunk@66 47920889-6d0e-0410-b696-d5024e80b239

+++ src/pysilc_callbacks.c
@@ -1032,7 +1032,7 @@ static void _pysilc_client_callback_ask_passphrase(SilcClient client,
     PYSILC_GET_CLIENT_OR_DIE(client, pyclient);
     PyObject *callback = NULL, *result = NULL;
     char *passphrase;
-	int length;
+    Py_ssize_t length;
 
     callback = PyObject_GetAttrString((PyObject *)pyclient, "ask_passphrase");
